Finalização do CRUD

+++ lib/revista.rb
@@ -6,16 +6,31 @@ class Revista
     @titulo = titulo
     @valor = valor
     @id = next_id
+    @destroyed = false
+    @new_record = true
   end
 
   def save
+    @new_record = false
     File.open("db/revistas/#{@id}.yml", 'w') do |file|
       file.puts serialize
     end
   end
 
   def self.find(id)
+    unless File.exist?("db/revistas/#{id}.yml")
+      raise DocumentNotFound, "Arquivo #{id} não encontrado.", caller
+    end
     YAML.load File.open("db/revistas/#{id}.yml", 'r')
+  end
+
+  def destroy
+    # verificar se o objeto nao foi removido anteriormente
+    # e se o objeto nao e um novo registro
+    unless @destroyed || @new_record
+      @destroyed = true
+      FileUtils.rm "db/revistas/#{@id}.yml"
+    end
   end
 
   private
